project/requirements.txt
fastapi==0.103.0
python-dotenv==1.0.0
uvicorn==0.23.2
black==23.7.0
pydantic-settings==2.0.3
asyncpg==0.28.0
SQLAlchemy==2.0.20
alembic==1.12.0
